+++ packages/network/lib/src/exception.dart
@@ -1,0 +1,8 @@
+import 'package:network/src/response.dart';
+
+class NetworkException<T extends BinaryResponse> implements Exception {
+  final T response;
+  NetworkException(this.response);
+
+  int get code => response.statusCode;
+}