+++ packages/extension/CHANGELOG.md
@@ -1,3 +1,14 @@
+[comment]: <> (Changelog bum example)
+[comment]: <> (## version)
+[comment]: <> (### Breaking Changes or ### New Features)
+[comment]: <> (* Change description)
+
+## 0.0.2
+
+### Breaking Changes
+
+* `enumValueByString` now throws `EnumValueNotFoundException` if can't find the value
+
 ## 0.0.1
 
 * Initial commit